Isolate ICB management from rest of the indexing
Move ICB map management into a new translation unit.

Create a new unit that picks the best match for the requested
index. Note that `find.c` will still call the indexing subsystem
multiple times to get several ICBs to choose from. Picking will
only change which ICBs will end up into that set of ICBs to select
from.
