+++ server/modules/selva/module/find_index/icb.c
@@ -1,0 +1,117 @@
+/*
+ * Copyright (c) 2021-2022 SAULX
+ * SPDX-License-Identifier: (MIT WITH selva-exception) OR AGPL-3.0-only
+ */
+#include <stddef.h>
+#include "base64.h"
+#include "redismodule.h"
+#include "svector.h"
+#include "selva.h"
+#include "selva_set.h"
+#include "traversal.h"
+#include "traversal_order.h"
+#include "hierarchy.h"
+#include "icb.h"
+
+size_t SelvaFindIndexICB_CalcNameLen(const Selva_NodeId node_id, const struct icb_descriptor *desc) {
+    size_t filter_len;
+    size_t n;
+
+    (void)RedisModule_StringPtrLen(desc->filter, &filter_len);
+
+    n = Selva_NodeIdLen(node_id) + base64_out_len(filter_len, 0) + 3;
+
+    if (desc->dir == SELVA_HIERARCHY_TRAVERSAL_BFS_EXPRESSION) {
+        size_t dir_expression_len;
+
+        (void)RedisModule_StringPtrLen(desc->dir_expression, &dir_expression_len);
+
+        /*
+         * Currently only expressions are supported in addition to fixed
+         * field name traversals.
+         */
+        n += base64_out_len(dir_expression_len, 0) + 1;
+    }
+
+    if (desc->sort.order != SELVA_RESULT_ORDER_NONE) {
+        size_t order_field_len;
+
+        (void)RedisModule_StringPtrLen(desc->sort.order_field, &order_field_len);
+
+        n += base64_out_len(order_field_len, 0) + 3;
+    }
+
+    return n;
+}
+
+void SelvaFindIndexICB_BuildName(char *buf, const Selva_NodeId node_id, const struct icb_descriptor *desc) {
+    size_t filter_len;
+    const char *filter_str = RedisModule_StringPtrLen(desc->filter, &filter_len);
+    char *s = buf;
+
+    /* node_id */
+    memcpy(s, node_id, Selva_NodeIdLen(node_id));
+    s += Selva_NodeIdLen(node_id);
+
+    /* direction */
+    *s++ = '.';
+    *s++ = 'A' + (char)__builtin_ffs(desc->dir);
+    *s++ = '.';
+
+    if (desc->dir == SELVA_HIERARCHY_TRAVERSAL_BFS_EXPRESSION) {
+        size_t dir_expression_len;
+        const char *dir_expression_str = RedisModule_StringPtrLen(desc->dir_expression, &dir_expression_len);
+
+        if (dir_expression_len > 0) {
+            s += base64_encode_s(s, dir_expression_str, dir_expression_len, 0);
+            *s++ = '.';
+        }
+    }
+
+    /* order */
+    if (desc->sort.order != SELVA_RESULT_ORDER_NONE) {
+        size_t order_field_len;
+        const char *order_field_str = RedisModule_StringPtrLen(desc->sort.order_field, &order_field_len);
+
+        *s++ = 'A' + (char)desc->sort.order;
+        *s++ = '.';
+        s += base64_encode_s(s, order_field_str, order_field_len, 0); /* This must be always longer than 0 */
+        *s++ = '.';
+    }
+
+    /* indexing clause filter */
+    s += base64_encode_s(s, filter_str, filter_len, 0);
+}
+
+int SelvaFindIndexICB_Get(struct SelvaHierarchy *hierarchy, const char *name_str, size_t name_len, struct SelvaFindIndexControlBlock **icb) {
+    struct SelvaObject *dyn_index = hierarchy->dyn_index.index_map;
+    void *p;
+    int err;
+
+    err = SelvaObject_GetPointerStr(dyn_index, name_str, name_len, &p);
+    *icb = p;
+
+    /* TODO This requires testing. Should we do it also here or only here? */
+#if 0
+    /*
+     * Increment popularity counter.
+     * We do it here too so that the counter is updated even if we end up
+     * falling back to another index.
+     */
+    if (icb && icb->pop_count.cur < INT_MAX) {
+        icb->pop_count.cur++;
+    }
+#endif
+
+    return err;
+}
+
+int SelvaFindIndexICB_Set(struct SelvaHierarchy *hierarchy, const char *name_str, size_t name_len, struct SelvaFindIndexControlBlock *icb) {
+    struct SelvaObject *dyn_index = hierarchy->dyn_index.index_map;
+
+    return SelvaObject_SetPointerStr(dyn_index, name_str, name_len, icb, NULL);
+}
+
+int SelvaFindIndexICB_Del(struct SelvaHierarchy *hierarchy, const struct SelvaFindIndexControlBlock *icb) {
+    return SelvaObject_DelKeyStr(hierarchy->dyn_index.index_map, icb->name_str, icb->name_len);
+}
