+++ server/modules/selva/module/find_index/pick_icb.h
@@ -1,0 +1,15 @@
+/*
+ * Copyright (c) 2022 SAULX
+ * SPDX-License-Identifier: (MIT WITH selva-exception) OR AGPL-3.0-only
+ */
+#pragma once
+#ifndef _FIND_INDEX_PICK_ICB_
+#define _FIND_INDEX_PICK_ICB_
+
+struct SelvaFindIndexControlBlock *SelvaFindIndexICB_Pick(
+        struct SelvaHierarchy *hierarchy,
+        const Selva_NodeId node_id,
+        const struct icb_descriptor *desc,
+        struct SelvaFindIndexControlBlock *first);
+
+#endif /* _FIND_INDEX_PICK_ICB_ */
